Separate Transaction data block on the transactions page into a separate component.
